bouncing che va

+++ src/App/UBOs.h
@@ -1,5 +1,6 @@
 #pragma once
+#include <cstdint>
 
 struct RandomSeed {
-    float seed;
+    uint32_t seed;
 };
